Added Files of Written Algorithm
These files contain the algorithm of which I wrote out as well they contain the proof by induction using all 4 cases where the target is greater or lesser than the current element as well where t both exists and doesn't exist. Also included is the time complexity analysis for this algorithm in which I use case 2 of masters theorem to prove this algorithm is theta of log_2(n).